Use pattern matching
Frequently we will attempt to parse the next token, expression, or statement and then check if the result is null on the next line. If we have succeeded, we continue parsing in a nested scope. Parsing an if/else or a variable declaration can get lengthy (with a lot of repetition) and deeply nested.

We can greatly improve this by switching to pattern matching, which replaces the variable assignment and null check with a single expression (and thus a single line). This in turn allows us to eliminate many nested scopes.