Support multi-line input
Right now the REPL reads a line of input and attemps to lex, parse, and execute it as-is. This means that the line must be a complete expression or statement, and there is no way to split entries across multiple lines (even though the lexer and parser don't care about lines).

This commit adds support for multi-line submissions.

We start with an empty input buffer, and read a single line of raw input. If the raw input is empty or null we take that to mean the user wants to reset their current submission, so we clear the input buffer and re-print the prompt. If it is _not_ empty we append the raw input to the input buffer. We then take the whole contents of the input buffer and attempt to parse it first as an expression and then as a statement. If either of these parses successfully we transform and execute the submission, print the result to the screen, and clear the input buffer. If neither parses successfully we assume the user is going to enter more on a further line, and return them to the prompt _without_ clearing the input buffer.

Also, we now vary the prompt. If the input buffer is empty we output '> ' to indicate they are starting a new submission. Otherwise we output '. ' to indicate that they are continuing a previous submission.